Move Preamble to 'yaml' package
It seems to fit this package as it deals with YAML emitting, where as
I consider DSL parts as stuff related to building the Actions workflows.
